Move video stretching object-fit transform workaround to mixins JW8-772
